chore: add code-split for fixing package version change bug with in-memory Git (#41442)
## Description

EE Counterpart: https://github.com/appsmithorg/appsmith-ee/pull/8468

Fixes #`Issue Number`
_or_
Fixes `Issue URL`
> [!WARNING]
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags="@tag.Git"

### :mag: Cypress test results
